examples/reliability/outbox_pattern/requirements.txt
azure-functions
azure-cosmos>=4.5.1
azure-functions-db
azure-functions-logging
